split timer and platform
