+++ Assets/Scripts/MovingElements/TimerTriangles.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ab507eb86f08c4b77af148bf08ccfbbe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
